requirements.txt
ebooklib
beautifulsoup4
google-generativeai
Pillow
